Do not leak internal API into public packages
Currently all scopes are leaking the internal package due to the use of
the (internal) AbstractScope class what is also used in platform itself
to implement ProjectScope.

This now makes this abstract class public and makes the old class
deprecated for removal with the next release (as it is only an internal
class) in case there are other references outside the platform.
