+++ bundles/org.eclipse.equinox.preferences/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: %pluginName
 Bundle-SymbolicName: org.eclipse.equinox.preferences; singleton:=true
-Bundle-Version: 3.11.500.qualifier
+Bundle-Version: 3.12.0.qualifier
 Bundle-Activator: org.eclipse.core.internal.preferences.Activator
 Bundle-Vendor: %providerName
 Bundle-Localization: plugin
@@ -11,7 +11,7 @@ Require-Bundle: org.eclipse.equinox.common;bundle-version="[3.18.0,4.0.0)",
  org.osgi.service.prefs;bundle-version="[1.1.0,1.2.0)";visibility:=reexport
 Export-Package: org.eclipse.core.internal.preferences;x-friends:="org.eclipse.core.resources,org.eclipse.core.runtime,org.eclipse.equinox.p2.engine",
  org.eclipse.core.internal.preferences.exchange;x-friends:="org.eclipse.core.runtime",
- org.eclipse.core.runtime.preferences;version="3.5.0";uses:="org.eclipse.core.runtime,org.osgi.service.prefs,org.eclipse.core.internal.preferences"
+ org.eclipse.core.runtime.preferences;version="3.6.0";uses:="org.eclipse.core.runtime,org.osgi.service.prefs"
 Bundle-ActivationPolicy: lazy; exclude:="org.eclipse.core.internal.preferences.exchange"
 Import-Package: org.eclipse.osgi.framework.log,
  org.eclipse.osgi.service.datalocation,
