+++ bundles/org.eclipse.equinox.preferences/src/org/eclipse/core/runtime/preferences/UserScope.java
@@ -13,7 +13,6 @@
  *******************************************************************************/
 package org.eclipse.core.runtime.preferences;
 
-import org.eclipse.core.internal.preferences.AbstractScope;
 import org.eclipse.core.runtime.IPath;
 
 /**
